backend/pill-identification/requirements.txt
# Pill Identification Module Requirements
# Based on ePillID-benchmark architecture

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pillow>=9.0.0

# Vector search
faiss-cpu>=1.7.4
# For GPU support, use: faiss-gpu>=1.7.4

# API framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Data processing
pandas>=1.5.0

# Utilities
tqdm>=4.65.0







